Issue: When there's an imageOverlay in the map and the map also has any other elements like polygons, circles, markers or even another imageOverlay, the images are deliberately ignored when forming the canvas.
Cause: When there's and imageOverlay in the map along with other elements, the map._panes.ovelayPane has as children more than one element, the canvas with all the elements and each imageOverlay included as independent child (IMG) of the overlayPane. The current code is only getting the map._pathRoot that is the canvas with all the elements, and this canvas does not include images, when map._pathRoot is not defined, the code only gets the first child of the overlayPan, ignoring any image or element added as child of it after the first element.

Fix: Changed if/else block with a for loop through all children of overlayPane to be drawn in the main canvas, there's no need to manage the difference between map._pathRoot and map._panes.overlayPane, because always the same element of the _pathRoot (a canvas) is included as a child of the map._panes.ovelayPane